state-res: Move event_auth tests to separate file
